requirements.txt
Flask==2.0.1
psycopg2-binary==2.9.1
pdfminer.six==20201018
openai==0.27.0
Werkzeug==2.0.1
